Remove call to `ratchet.py` and `--workspace` from clank (#71)
* Make ratchet executable

* Rename FILE_PATH to CUR_DIR

* Clean up arguments

* Remove workspace

This could have been split up a little better.

Move env_file with other required options above
```diff
-    # Load env file
-    if args.env_file:
-        command += ' -e "@%s/clank/%s"' % (workspace, args.env_file)
```

Add inventory option like ansible rather than expecting clank/local_inventory
```diff
+    parser.add_argument("-i", "--inventory",
+       default="{}/hosts".format(CUR_DIR),
+        type=str,
+        help="The path to the ansible inventory, which defaults to ./hosts.")
```

* Keep main responsible for everything

* Catch missing imports, Remove old requirements

* Remove ratchet, only clank

I think the ratchet/clank distinction is going to add confusion as more people
(outside) are using this tool.

* Improve options, correct spelling

The following change:
```diff
     parser.add_argument("--skip-tags",
	...
+       metavar="TAGS",
	...
```

Results in `"TAGS"` appearing in the help output:
```diff
- --skip-tags SKIP-TAGS
+ --skip-tags TAGS
```

* Reword README

Prefer including to skipping, since skipping requires you to know
everything that might run. There is another replacement of Ratchet.py to
clank.py.

* Revert "Remove workspace"

This reverts commit ad255047039967cdd73d907541f37c7ef3c87f65.

* Require virtual env, improve errors

* Remove extra files/installs

* Improve options

* Prefer memory over redis

* Provide install from original dependencies

* Remove unused import

* Make --help not require virtualenv

* Remove logging by default

* Prefer local to global vars

* Improve --help output
