+++ ansible.cfg
@@ -57,7 +57,7 @@ timeout = 10
 
 # logging is off by default unless this path is defined
 # if so defined, consider logrotate
-log_path = /var/log/ansible/ansible.log
+# log_path = /var/log/ansible/ansible.log
 
 # default module name for /usr/bin/ansible
 #module_name = command
@@ -150,7 +150,7 @@ filter_plugins     = /usr/share/ansible_plugins/filter_plugins:./filter_plugins
 # wanting to use, for example, IP information from one group of servers
 # without having to talk to them in the same playbook run to get their
 # current IP information.
-fact_caching = redis
+fact_caching = memory
 # 24 hours of fact caching
 fact_caching_timeout = 86400
