--- create_virtualenv.txt
@@ -1,3 +1,0 @@
-pip install --upgrade pip virtualenv virtualenvwrapper
-virtualenv clank_env
-clank_env/bin/pip install -r clank_requirements.txt
