--- install_dependencies.txt
@@ -1,11 +1,0 @@
-apt-get update
-apt-get purge redis-tools -y
-apt-get install -y software-properties-common
-apt-get install -y python-pip python-dev build-essential
-apt-add-repository ppa:ansible/ansible
-add-apt-repository ppa:chris-lea/redis-server
-apt-get update
-apt-get install -y ansible 
-apt-get install -y redis-server
-mkdir -p /var/log/ansible
-touch /var/log/ansible/ansible.log
