--- local_inventory
@@ -1,2 +1,0 @@
-[all]
-localhost
