--- ratchet_requirements.txt
@@ -1,6 +1,0 @@
-envoy
-MarkupSafe
-Jinja2
-PyYAML 
-ruamel.yaml
-colorama
