+++ requirements.txt
@@ -1,2 +1,2 @@
+colorama
 ansible==2.1.0.0
-redis
